Removed home screen names and renamed it to Welcome screen
