--- app/ts/HomeScreen/HomeScreen.tsx
@@ -1,28 +1,0 @@
-/* eslint-disable prettier/prettier */
-/**
- * @providesModule HomeScreen
- * @flow
- */
-
-import React from 'react';
-import styles from './HomeScreenStyle';
-
-import {
-  View,
-} from 'react-native';
-
-export default class HomeScreen extends React.Component<{}, {}> {
-
-  constructor(props: any) {
-    super(props);
-  }
-
-
-  render(): React.ReactElement {
-    return (
-      <View>
-
-      </View>
-    );
-  }
-}
