--- app/ts/HomeScreen/HomeScreenStyle.ts
@@ -1,12 +1,0 @@
-/* eslint-disable prettier/prettier */
-/**
- * @provideModule HomeScreenStyle
- * @flow
- */
-
-import { StyleSheet } from 'react-native';
-
-
-export default StyleSheet.create({
-  
-});
